+++ app/src/main/java/com/sanaa/tudee_assistant/TudeeAssistantApplication.kt
@@ -21,7 +21,7 @@ class TudeeAssistantApplication : Application() {
                 databaseModule,
                 servicesModule,
                 imageProcessorModule,
-                viewModelModule
+                themeManagerModule,
             )
         }
     }
