+++ app/src/main/java/com/sanaa/tudee_assistant/di/viewModelModule.kt
@@ -1,10 +1,10 @@
 package com.sanaa.tudee_assistant.di
 
-import com.sanaa.tudee_assistant.data.utils.ImageProcessor
+import com.sanaa.tudee_assistant.domain.ImageProcessor
 import com.sanaa.tudee_assistant.domain.service.CategoryService
 import com.sanaa.tudee_assistant.domain.service.TaskService
-import com.sanaa.tudee_assistant.presentation.screens.category.CategoryViewModel
 import com.sanaa.tudee_assistant.presentation.screen.home.HomeScreenViewModel
+import com.sanaa.tudee_assistant.presentation.screens.category.CategoryViewModel
 import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
