+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/screens/category/CategoryScreen.kt
@@ -37,7 +37,7 @@ import com.sanaa.tudee_assistant.R
 import com.sanaa.tudee_assistant.presentation.design_system.component.CategoryCount
 import com.sanaa.tudee_assistant.presentation.design_system.component.CategoryItem
 import com.sanaa.tudee_assistant.presentation.design_system.theme.Theme
-import com.sanaa.tudee_assistant.presentation.model.CategoryState
+import com.sanaa.tudee_assistant.presentation.state.CategoryState
 import org.koin.androidx.compose.koinViewModel
 
 @OptIn(ExperimentalMaterial3Api::class)
